LayoutTests/storage/indexeddb/constants-expected.txt

Test IndexedDB's constants.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

PASS IDBDatabaseException.UNKNOWN_ERR is 1
PASS IDBDatabaseException.NON_TRANSIENT_ERR is 2
PASS IDBDatabaseException.NOT_FOUND_ERR is 3
PASS IDBDatabaseException.CONSTRAINT_ERR is 4
PASS IDBDatabaseException.DATA_ERR is 5
PASS IDBDatabaseException.NOT_ALLOWED_ERR is 6
PASS IDBDatabaseException.TRANSACTION_INACTIVE_ERR is 7
PASS IDBDatabaseException.ABORT_ERR is 8
PASS IDBDatabaseException.READ_ONLY_ERR is 9
PASS IDBDatabaseException.TIMEOUT_ERR is 10
PASS IDBDatabaseException.QUOTA_ERR is 11
PASS IDBDatabaseException.VER_ERR is 12
PASS IDBRequest.LOADING is 1
PASS IDBRequest.DONE is 2
PASS IDBCursor.NEXT is 0
PASS IDBCursor.NEXT_NO_DUPLICATE is 1
PASS IDBCursor.PREV is 2
PASS IDBCursor.PREV_NO_DUPLICATE is 3
PASS IDBTransaction.READ_ONLY is 0
PASS IDBTransaction.READ_WRITE is 1
PASS IDBTransaction.VERSION_CHANGE is 2
PASS successfullyParsed is true

TEST COMPLETE
